indexnow/requirements.txt
requests==2.28.1
google-auth==2.23.0
google-auth-oauthlib==1.2.0